+++ client/src/components/Entrance.tsx
@@ -2,7 +2,7 @@ import { useState } from "react";
 
 import { Login } from "./Login";
 import { Register } from "./Register";
-import { Shared } from "./Shared";
+import { Topbar } from "./Topbar";
 
 export function Entrance({ userRequest }: basicprops): JSX.Element {
 	const [isLogin, setIsLogin] = useState<boolean>(true);
@@ -10,7 +10,7 @@ export function Entrance({ userRequest }: basicprops): JSX.Element {
 	return (
 		<div className="entrance-wrapper">
 			<div className="entrance">
-				<Shared showLogout={false} userRequest={userRequest} />
+				<Topbar showLogout={false} userRequest={userRequest} />
 
 				{(isLogin)
 					? <Login setIsLogin={setIsLogin} userRequest={userRequest} />