+++ client/src/shared/capitalizeFirstLetter.ts
@@ -1,4 +1,4 @@
-export function capitalize(str: string) {
+export function capitalizeFirstLetter(str: string) {
 	if (typeof str !== "string") return "";
 	return str.charAt(0).toUpperCase() + str.slice(1);
 }