+++ server/src/middleware/database.ts
@@ -16,7 +16,7 @@ export async function connect(): Promise<void> {
 			password VARCHAR(100) UNIQUE NOT NULL,
 			created TIMESTAMP DEFAULT NULL
 		)`)
-		.then((res: QueryResult<any>) => { if (!PRODUCTION) { output("Table 'users' created.", "yellow"); } })
+		.then(() => { if (!PRODUCTION) { output("Table 'users' created.", "yellow"); } })
 		.catch((err: Error) => { throw console.log(err); });
 
 	client.query(
@@ -32,7 +32,7 @@ export async function connect(): Promise<void> {
 			discord_channel VARCHAR,
 			created TIMESTAMP DEFAULT NULL
 		)`)
-		.then((res: QueryResult<any>) => { if (!PRODUCTION) { output("Table 'campaigns' created.", "yellow"); } })
+		.then(() => { if (!PRODUCTION) { output("Table 'campaigns' created.", "yellow"); } })
 		.catch((err: Error) => { throw console.log(err); });
 
 	client.query(
@@ -46,7 +46,7 @@ export async function connect(): Promise<void> {
 			campaign_name VARCHAR(32),
 			created TIMESTAMP DEFAULT NULL
 		)`)
-		.then((res: QueryResult<any>) => { if (!PRODUCTION) { output("Table 'characters' created.", "yellow"); } })
+		.then(() => { if (!PRODUCTION) { output("Table 'characters' created.", "yellow"); } })
 		.catch((err: Error) => { throw console.log(err); });
 
 	client.release();